requirements.txt
codecarbon
transformers
pillow
datasets
--extra-index-url https://download.pytorch.org/whl/cpu
torch
